requirements.txt
asgiref==3.8.1
boto3==1.39.0
botocore==1.39.0
Django==5.2.1
django-environ==0.12.0
django-storages==1.14.6
et_xmlfile==2.0.0
gunicorn==23.0.0
jmespath==1.0.1
openpyxl==3.1.5
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
s3transfer==0.13.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.9.0
gunicorn>=20.0